+++ src/libs/productService.ts
@@ -1,5 +1,5 @@
 import axios from "axios";
-import { v4 as uuidv4 } from "uuid";
+// import { v4 as uuidv4 } from "uuid";
 
 // export const ProductSchema = Yup.object({
 //   id: Yup.string(),
